+++ src/md/Algorithm.mdx
@@ -35,7 +35,7 @@
 5. `효과성`
 
 ### 알고리즘의 수행시간
-![image](https://user-images.githubusercontent.com/45661217/146776606-2fe2153e-2f60-495f-93e2-51dff85df636.png)  
+
 O(Big-o)표기법 : 자료에 따라 알고리즘이 수행되는 최악의 경우의 성능을 표시하는 방법
 
 > 알고리즘 수행 세는 방법
