+++ src/md/Cors.mdx
@@ -9,3 +9,5 @@
     - 크롬에서 –disable-web-security 옵션을 추가
 - Access-Control-Allow-Origin 헤더는 뭘까?
 - cors를 사용하는 이유
+
+> 1월달 즈음에 업로드 예정입니다.
