+++ src/md/Flask.mdx
@@ -1,4 +1,8 @@
 # FLASK 튜토리얼
+
+> 개발 테스트용 포스트입니다. 1월달 즈음 업데이트 될 예정입니다.
+> flask 튜토리얼이 아니며 github에 간략하게 썼던 내용을 가져왔습니다.
+
 > <a href="https://github.com/min050410/Flask-take-notes/tree/master">View my code</a>    
 
 ### 개발시에 시간이 오래 걸렸거나 헷갈렸던 것들 모임
