+++ src/md/recent.tsx
@@ -11,7 +11,7 @@ const recent = [
       title: 'Big-O 표기법과 수행',
       filename: 'Algorithm',
       img: 'algorithm',
-      date: '12월 20일',
+      date: '2021-12-20',
       tag: 'Big O',
       tag2: 'sort',
       tag3: 'DP',
@@ -21,7 +21,7 @@ const recent = [
       title: 'Window 터미널 꾸미기',
       filename: 'Ohmyposh',
       img: 'ohmyposh',
-      date: '12월 18일',
+      date: '2021-12-18',
       tag: 'terminal',
       tag2: 'window',
       tag3: '',
@@ -31,7 +31,7 @@ const recent = [
       title: '비선형구조 정리',
       filename: 'Structure',
       img: 'datastucture',
-      date: '12월 17일',
+      date: '2021-12-17',
       tag: 'C',
       tag2: 'tree',
       tag3: 'graph',
@@ -41,7 +41,7 @@ const recent = [
       title: 'Flask 튜토리얼',
       filename: 'Flask',
       img: 'flask',
-      date: '12월 11일',
+      date: '2021-12-11',
       tag: 'python',
       tag2: 'flask',
       tag3: 'backend',
@@ -51,7 +51,7 @@ const recent = [
       title: 'CORS와 해결방법',
       filename: 'Cors',
       img: 'CORS',
-      date: '12월 9일',
+      date: '2021-12-09',
       tag: 'cors',
       tag2: 'frontend',
       tag3: 'backend',
@@ -60,7 +60,7 @@ const recent = [
       id: 6,
       title: 'RN camera 적용법',
       filename: 'Camera',
-      date: '11월 26일',
+      date: '2021-11-26',
       img: 'camera',
       tag: 'react-native',
       tag2: 'module',
@@ -70,7 +70,7 @@ const recent = [
       id: 7,
       title: '타입스크립트를 많이 쓰는 이유',
       filename: 'Typescript',
-      date: '11월 17일',
+      date: '2021-11-17',
       img: 'typescript',
       tag: 'react',
       tag2: 'ts',
