+++ src/pages/login.tsx
@@ -12,5 +12,4 @@ const login = () => {
         }, [])
         return null;
 }
-
 export default login;
