+++ src/styles/postitem.sass
@@ -153,6 +153,8 @@ code
   pre
     padding: 0px
     font-size: 1rem 
+  .middle
+    margin: 0px;
